+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tweet relaying has been completely revamped into `feeds`. Feed relaying will need to be reconfigured.
 - A database migration is required.
 - The `TwitterOptions` config section has been renamed to `FeedsPluginOptions`, and the properties appended with `Twitter`.
+- The API application has been completed removed!
 
 ## Release v0.5.2